Updated to work with Leaflet 1.0; bumped version to 0.3.0
The L.Control.Layers class changed in Leaflet 1.0+ to store the
`_layers` object as a list rather than using an Object with
the `L.stamp()` (layerId) as keys.  This broke retrieval of overlays.

Old format:
  _layers : { 1: {layer: ..., name: ...}, 104: {layer: ..., name: ... } }

New format:
  _layers : [ {layer: ..., name: ... }, {layer: ..., name: ... } ]

The L.stamp() value can be retrieved using `layer._leaflet_id` in this
scenario, so the return value from getActiveOverlayLayers() has not
changed.

After this change, the plugin should be compatible with both the
Leaflet 0.7.x and 1.0 branches.